requirements.txt
faiss-cpu
pdfplumber
numpy
uvicorn
fastapi
langchain-text-splitters
sentence-transformers
torch
